Refactor of src/diff.rs
src/diff.rs can now be found in two parts:
src/diff/mod.rs holds the front facing part of the module.
src/diff/diff_type.rs holds the type parcing and state machine logic
